Fix identity module protobuf compilation errors
- Regenerate protobuf files for identity module
- Fix proto imports (pagination, genesis, module_query_safe)
- Update go_package to match project module
- Remove duplicate type declarations (now in pb.go files)
- Update field names to match proto (DID -> Did, int64 -> uint64)
- Fix sdk.Coin to string conversions in limit types
- Update msg_server.go and query_server.go to use context.Context
- Add sdk.Msg interface methods (GetSigners, ValidateBasic, etc.)
- Update module.go to use service-based architecture
- Fix handler.go error handling

+++ x/identity/keeper/limit.go
@@ -72,7 +72,7 @@ func (k Keeper) UpdateLimitConfig(ctx sdk.Context, targetAddress string, newConf
 	}
 
 	// Update timestamp
-	newConfig.UpdatedAt = ctx.BlockTime().Unix()
+	newConfig.UpdatedAt = uint64(ctx.BlockTime().Unix())
 
 	// Store the config
 	k.SetLimitConfig(ctx, newConfig)
